Add industrial benchmarks.
We are happy to release the enclosed SMTLIB benchmarks under the
GPL v3. They should represent the kind of real-world floating point
verification conditions for absence of run-time errors that
we have to deal with on safety critical code.

The benchmarks have been extracted from a subset of a large industrial
Ada 2005 code base. I used the SPARK 2014 tools to produce verification
conditions, and a python script to replace identifiers and filenames.
The sat/unsat status in each benchmark should be correct, but since
there may be issues with the encoding I can't fully guarantee it.