Allow per-benchmark suite timeouts
This is an alt-ergo specific hack that bounds run-time on the
wintersteiger suite as it time-outs on basically all SAT tests
which kills overall benchmark runtime as there are a lot of these.

However that testsuite contains no interesting (read "hard") benchmarks
so no solver is actually impacted by the lower timeout here.