Add script to port benchmarks from the old DT format to SMTLIB 2.6
Apply this script to industrial_1. We use CVC4 to do this translation, but
it seems to do some rewriting (adding lets) that I just cannot get rid of.

Eventually I will need to finish PySMT and use that to re-write this
with preserving the exact form.

This script doesn't work in general: there seems to be some issues in CVC4
when it comes to dumping benchmarks.
See https://github.com/CVC4/CVC4/issues/3005